requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
distro==1.9.0
Django==5.1
django-cors-headers==4.4.0
django-environ==0.11.2
django-filter==24.3
django-ratelimit==4.1.0
django-rest-passwordreset==1.4.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
jiter==0.5.0
numpy==2.1.2
openai==1.44.0
openpyxl==3.1.5
pandas==2.2.3
pydantic==2.9.0
pydantic_core==2.23.2
PyJWT==2.9.0
pyspellchecker==0.8.1
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
xlrd==2.0.1
